requirements.txt
sentence-transformers
numpy
tree-sitter
tree-sitter-python
tree-sitter-kotlin
tree-sitter-go
ollama
argparse
tqdm
pydantic
pyyaml
# onnxruntime
